chore: move `EcdsaConfigWithInstance` to chip folder
