+++ zk_prover/src/chips/mod.rs
@@ -1,4 +1,5 @@
 pub mod aggregation;
+pub mod ecdsa;
 pub mod merkle_sum_tree;
 pub mod overflow;
 pub mod poseidon;
